projects/cloudcurio-kb/services/bff/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.9.2
python-jose[cryptography]==3.3.0
cachetools==5.5.0
loguru==0.7.2
opensearch-py==2.6.0
qdrant-client==1.10.1
psycopg[binary]==3.2.1
pyyaml==6.0.2
